Provide indent contexts that may override indent information to indent helpers
So that region-indent can properly reindent even if reference lines
for some of the lines below are reindented in the same operation.

BREAKING: The interface to `EditorState.indentation` changed, taking
an `IndentContext` object.

FEATURE: It is now possible to provide indentation functions with information
about already-reindented lines through an `IndentContext` object, making it
possible to reindent a region in one go without re-parsing.

Closes #162