Add isUserEvent and TransactionSpec.userEvent
FEATURE: It is now possible to add a user event annotation with a direct
`userEvent` property on a transaction spec.

FEATURE: Transactions now have an `isUserEvent` method that can be used
to check if it is (a subtype of) some user event type.

BREAKING: User event strings now work differently—the events emitted by
the core packages follow a different system, and hierarchical event tags
can be created by separating the words with dots.